+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v3.2.1-1.18.2] - 2023-02-17
+### Added
+- Added dimension specific blacklists, also supports inverting the lists to be used as whitelists
+- Check the `README` on [GitHub](https://github.com/Fuzss/pickupnotifier) for more information!
+
 ## [v3.2.0-1.18.2] - 2022-03-03
 - Compiled for Minecraft 1.18.2